eda_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
aiohttp==3.9.1
aiofiles==23.2.1
minio==7.2.0
# ydata-profiling 4.6.x: pandas < 2.1; numpy >=1.16 и <1.26. pandas 2.0.3 на py3.11: numpy >=1.23.2
pandas==2.0.3
numpy==1.25.2
ydata-profiling==4.6.1
python-jose[cryptography]==3.3.0
certifi==2023.11.17
